+++ app/components/SliderEntry.js
