+++ app/resources/beforeClock.png
