+++ app/resources/nowClock.png
