+++ app/resources/stomachPain.png
