+++ app/styles/SliderEntry.style.js
