+++ package.json
@@ -28,7 +28,6 @@
     "react-native-extended-stylesheet": "^0.8.1",
     "react-native-snap-carousel": "^3.6.0",
     "react-native-swipeout": "^2.3.3",
-    "react-native-two-way-list": "^1.1.0",
     "react-redux": "^5.0.6",
     "redux-logger": "^3.0.6"
   }
